Extract onboarding view module
