+++ js/onboarding-view.js
@@ -1,0 +1,158 @@
+// onboarding-view.js - Dashboard onboarding and AI connection reminders
+
+import { state } from './state.js';
+import { getActiveData, updateHeaderDates } from './data.js';
+import { profileStorageKey, setProfileSex, setProfileDob } from './profile.js';
+import { hasAIProvider } from './api.js';
+import { showNotification } from './utils.js';
+
+let _navigate = null;
+
+export function configureOnboardingView(options = {}) {
+  _navigate = typeof options.navigate === 'function' ? options.navigate : null;
+}
+
+export function renderOnboardingBanner() {
+  const onboarded = localStorage.getItem(profileStorageKey(state.currentProfile, 'onboarded'));
+  if (onboarded) return '';
+  if (state.profileSex && state.profileDob) {
+    localStorage.setItem(profileStorageKey(state.currentProfile, 'onboarded'), 'profile-set');
+    return '';
+  }
+  return `<div class="onboarding-banner" id="onboarding-banner">
+    <div class="onboarding-steps">
+      <span class="onboarding-step completed">\u2713</span>
+      <span class="onboarding-step-line"></span>
+      <span class="onboarding-step active">2</span>
+      <span class="onboarding-step-line"></span>
+      <span class="onboarding-step">3</span>
+    </div>
+    <div class="onboarding-step-labels">
+      <span class="onboarding-step-label">Import</span>
+      <span class="onboarding-step-label active">Profile</span>
+      <span class="onboarding-step-label">Ready</span>
+    </div>
+    <h3 class="onboarding-title">Set up your profile</h3>
+    <p class="onboarding-subtitle">Sex and date of birth pick the right reference ranges for your results.</p>
+    <div class="onboarding-form">
+      <div class="onboarding-field">
+        <label class="onboarding-label">Sex</label>
+        <div class="onboarding-sex-toggle">
+          <button class="onboarding-sex-btn${state.profileSex === 'male' ? ' active' : ''}" onclick="completeOnboardingSex('male')">Male</button>
+          <button class="onboarding-sex-btn${state.profileSex === 'female' ? ' active' : ''}" onclick="completeOnboardingSex('female')">Female</button>
+        </div>
+      </div>
+      <div class="onboarding-field">
+        <label class="onboarding-label">Date of Birth</label>
+        <input type="date" class="onboarding-dob-input" id="onboarding-dob" value="${state.profileDob || ''}" />
+      </div>
+      <div class="onboarding-actions">
+        <button class="onboarding-save-btn" onclick="completeOnboardingProfile()">Save & Continue</button>
+        <button class="onboarding-skip-btn" onclick="dismissOnboarding()">Skip for now</button>
+      </div>
+    </div>
+  </div>`;
+}
+
+export function completeOnboardingSex(sex) {
+  document.querySelectorAll('.onboarding-sex-btn').forEach(b => b.classList.remove('active'));
+  const btns = document.querySelectorAll('.onboarding-sex-btn');
+  if (sex === 'male' && btns[0]) btns[0].classList.add('active');
+  if (sex === 'female' && btns[1]) btns[1].classList.add('active');
+}
+
+export function completeOnboardingProfile() {
+  const activeSexBtn = document.querySelector('.onboarding-sex-btn.active');
+  const sex = activeSexBtn ? (activeSexBtn.textContent.trim().toLowerCase()) : null;
+  const dobInput = document.getElementById('onboarding-dob');
+  const dob = dobInput ? dobInput.value : null;
+  localStorage.setItem(profileStorageKey(state.currentProfile, 'onboarded'), 'profile-set');
+  if (sex) { state.profileSex = sex; setProfileSex(state.currentProfile, sex); }
+  if (dob) { state.profileDob = dob; setProfileDob(state.currentProfile, dob); }
+  const data = getActiveData();
+  window.buildSidebar(data);
+  updateHeaderDates(data);
+  (_navigate || window.navigate)?.('dashboard', data);
+  showNotification("Profile set up \u2014 you're all set!", 'success');
+}
+
+export function dismissOnboarding() {
+  localStorage.setItem(profileStorageKey(state.currentProfile, 'onboarded'), 'dismissed');
+  const banner = document.getElementById('onboarding-banner');
+  if (banner) {
+    banner.style.transition = 'opacity 0.3s, transform 0.3s';
+    banner.style.opacity = '0';
+    banner.style.transform = 'translateY(-10px)';
+    setTimeout(() => banner.remove(), 300);
+  }
+  showNotification('You can set sex and DOB anytime in Settings.', 'info');
+}
+
+export function renderAIConnectionReminder() {
+  if (hasAIProvider()) return '';
+  const skipKey = `labcharts-onboard-provider-skipped-${state.currentProfile}`;
+  const skipped = localStorage.getItem(skipKey);
+  if (!skipped) return '';
+  const dismissKey = profileStorageKey(state.currentProfile, 'ai-reminder-dismissed');
+  if (localStorage.getItem(dismissKey)) return '';
+  return `<div class="ai-reminder-banner" id="ai-reminder-banner" role="region" aria-label="Connect AI to unlock lab analysis">
+    <span class="ai-reminder-icon" aria-hidden="true">&#129504;</span>
+    <span class="ai-reminder-body">
+      <strong>Connect AI to unlock lab analysis</strong>
+      <span>PDF import, trend insights, and chat all need an AI provider. About 30 seconds.</span>
+    </span>
+    <button type="button" class="ai-reminder-cta" onclick="window.openChatProviderQuiz()">Connect now</button>
+    <button type="button" class="ai-reminder-dismiss" onclick="window.dismissAIReminder()" aria-label="Dismiss">&times;</button>
+  </div>`;
+}
+
+export function dismissAIReminder() {
+  const dismissKey = profileStorageKey(state.currentProfile, 'ai-reminder-dismissed');
+  localStorage.setItem(dismissKey, '1');
+  const banner = document.getElementById('ai-reminder-banner');
+  if (banner) {
+    banner.style.transition = 'opacity 0.3s, transform 0.3s';
+    banner.style.opacity = '0';
+    banner.style.transform = 'translateY(-10px)';
+    setTimeout(() => banner.remove(), 300);
+  }
+}
+
+export function setOnboardingFocus(mode) {
+  const body = document.body;
+  body.classList.remove('cards-focus', 'import-focus');
+  if (!mode) return;
+  if (mode === 'cards') {
+    body.classList.add('cards-focus');
+  } else if (mode === 'import') {
+    body.classList.add('import-focus');
+  }
+  if (body.classList.contains('chat-fullscreen')) {
+    body.classList.remove('chat-fullscreen');
+    localStorage.setItem('labcharts-chat-fullscreen', 'false');
+  }
+  if (mode === 'cards') {
+    const cards = document.querySelector('.profile-context-cards');
+    if (cards) {
+      setTimeout(() => cards.scrollIntoView({ behavior: 'smooth', block: 'start' }), 100);
+    } else if (window.openChatPanel) {
+      body.classList.remove('cards-focus');
+      const prefill = hasAIProvider()
+        ? 'Help me collect the health context you need before I import labs. Ask me one question at a time.'
+        : undefined;
+      window.openChatPanel(prefill);
+    }
+  } else if (mode === 'import') {
+    setTimeout(() => document.querySelector('.welcome-direct-import-btn, .welcome-primary-panel')?.scrollIntoView({ behavior: 'smooth', block: 'center' }), 100);
+  }
+}
+
+export function openChatProviderQuiz() {
+  const skipKey = `labcharts-onboard-provider-skipped-${state.currentProfile}`;
+  localStorage.removeItem(skipKey);
+  sessionStorage.setItem(`chat-onboard-provider-requested-${state.currentProfile}`, '1');
+  sessionStorage.removeItem(`chat-onboard-provider-branch-${state.currentProfile}`);
+  if (window.openChatPanel) window.openChatPanel();
+  else if (window.toggleChatPanel) window.toggleChatPanel();
+  if (window.renderChatMessages) window.renderChatMessages();
+}
