+++ service-worker.js
@@ -48,6 +48,7 @@ const APP_SHELL = [
   '/js/cycle.js',
   '/js/context-cards.js',
   '/js/focus-card.js',
+  '/js/onboarding-view.js',
   '/js/emf.js',
   '/js/image-utils.js',
   '/js/pdf-import.js',
